refactor: reduce dom.js coupling by extracting domain-specific helpers
Extract buildChevronRow to chevron-row.js (2 consumers only).
Create domain facade modules (terminal-dom, tab-dom, workspace-dom,
flow-dom, file-dom, settings-dom) so domain-specific files import
through their facade instead of directly from dom.js.

Direct dom.js importers reduced from 45 to 21 (7 are facades).
