+++ src/components/file-tree.js
@@ -1,4 +1,5 @@
-import { _el, createActionButton, buildChevronRow } from '../utils/dom.js';
+import { _el, createActionButton } from '../utils/file-dom.js';
+import { buildChevronRow } from '../utils/dom.js';
 import {
   CHEVRON_EXPANDED, CHEVRON_COLLAPSED,
   DEBOUNCE_DELAY, WATCH_PREFIX,
