+++ src/components/flow-card-terminal.js
@@ -2,7 +2,7 @@
  * Terminal management for flow cards: live terminals, inline log terminals, and log modal.
  * Extracted from FlowView to reduce component size.
  */
-import { _el } from '../utils/dom.js';
+import { _el } from '../utils/flow-dom.js';
 import { createModalOverlay } from '../utils/dom-dialogs.js';
 import { setupKeyboardShortcuts } from '../utils/keyboard-helpers.js';
 import { _safeFit, createReadonlyTerminal, disposeTerminal, disposeTerminalMap } from '../utils/terminal-factory.js';
