+++ src/utils/chevron-row.js
@@ -1,0 +1,9 @@
+/**
+ * Chevron + name row builder.
+ *
+ * Used by file-tree rows, flow-category headers, and tab elements — any
+ * place that needs the common "chevron + label" pattern.
+ *
+ * Re-exported from dom.js to reduce hub coupling.
+ */
+export { buildChevronRow } from './dom.js';
