+++ src/utils/file-tree-renderer.js
@@ -4,7 +4,8 @@
  */
 
 import { emitFileOpen } from './workspace-events.js';
-import { _el, buildChevronRow } from './dom.js';
+import { _el } from './file-dom.js';
+import { buildChevronRow } from './dom.js';
 import { computeIndent, CHEVRON_EXPANDED, CHEVRON_COLLAPSED, SVG_ICONS } from './file-tree-helpers.js';
 import { buildFileContextItems, buildDirContextItems } from './file-tree-context-menu.js';
 import { attachContextMenu } from './context-menu.js';
