+++ src/utils/settings-section-builder.js
@@ -3,7 +3,7 @@
  * Abstracts the heading + content + action buttons pattern
  * shared across settings-appearance, settings-configs, and settings-keybindings.
  */
-import { _el, renderButtonBar } from './dom.js';
+import { _el, renderButtonBar } from './settings-dom.js';
 import { createAsyncHandler } from './event-helpers.js';
 
 /**
